requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.51.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
tk==0.1.0
tzdata==2024.1
urllib3==2.2.1
